model/requirements.txt
numpy
opencv-python
face_recognition
cloudinary
fastapi
python-dotenv
requests
python-multipart
uvicorn